% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bdparData.R
\docType{data}
\name{bdparData}
\alias{bdparData}
\title{Example of the content of the files to be preprocessed.}
\format{A data frame with 20 rows and 2 variables:
\describe{
  \item{path}{File path.}
  \item{source}{File content.}
}}
\usage{
data(bdparData)
}
\description{
A manually collected data set containing e-mails and SMS messages from the
nutritional and health domain classified as spam and non-spam (with a ratio
of 50\%). In addition the dataset contains two variables: (i) path which
indicates the location of the target file and, (ii) source which contains
the raw text comprising each file.
}
\keyword{datasets}
