% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{busca_bdpa}
\alias{busca_bdpa}
\title{Busca BDPA}
\usage{
busca_bdpa(query)
}
\arguments{
\item{query}{Expressão de busca para recuperar dados na BDPA. Aceita operadores boleanos como "AND","OR", "NOT" e outros.}
}
\value{
dataframe with BDPA's data
}
\description{
Função para realizar busca na BDPA
}
\examples{
data <- busca_bdpa("pecuária AND mato grosso AND leiteira")
head(data)
}
