% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emojisData.R
\docType{data}
\name{emojisData}
\alias{emojisData}
\title{Emojis codes and descriptions data.}
\format{
A data frame with 2623 rows and 2 variables:
\describe{
  \item{code}{Emoji code}
  \item{description}{Emoji description.}
}
}
\usage{
data(emojisData)
}
\description{
This data comes from "Unicode.org",
<http://unicode.org/emoji/charts/full-emoji-list.html>. The data are codes
and descriptions of Emojis.
}
\keyword{datasets}
