% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenericPipeline.R
\docType{class}
\name{GenericPipeline}
\alias{GenericPipeline}
\title{Absctract super class implementing the pipelining proccess.}
\usage{
GenericPipeline
}
\description{
Abstract super class to establish the flow of Pipes.
}
\section{Constructor}{

\code{GenericPipeline$new()}
}

\section{Methods}{

\itemize{
\item{\bold{execute:}}{
function where is implemented the flow of the pipes.
\itemize{
\item{\emph{Usage:}}{
\code{execute(instance)}
}
\item{\emph{Value:}}{
the preprocessed \code{\link{Instance}}.
}
\item{\emph{Arguments:}}{
\itemize{
\item{\strong{instance:}}{
(\emph{Instance}) the \code{\link{Instance}} that is going to be processed.
}
}
}
}
}

\item{\bold{get:}}{
gets a list with containinig the set of pipes of the pipeline,
\itemize{
\item{\emph{Usage:}}{
\code{get()}
}
\item{\emph{Value:}}{
the set of pipes containing the pipeline.
}
}
}
}
}

\seealso{
\code{\link{DefaultPipeline}}, \code{\link{DynamicPipeline}},
         \code{\link{Instance}}, \code{\link{GenericPipe}}, \code{\link{\%>I\%}}
}
