% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InstanceFactory.R
\docType{class}
\name{InstanceFactory}
\alias{InstanceFactory}
\title{Class to handle the creation of Instance types}
\usage{
InstanceFactory
}
\description{
\code{\link{InstanceFactory}} class builds the appropriate
\code{\link{Instance}} object according to the file extension.
}
\section{Constructor}{

\code{InstanceFactory$new()}
}

\section{Methods}{

\itemize{
\item{\bold{createInstance:}}{
builds the \code{\link{Instance}} object according to the file extension.
\itemize{
\item{\emph{Usage:}}{
\code{createInstance(path)}
}
\item{\emph{Value:}}{
the \code{\link{Instance}} corresponding object according to the file extension.
}
\item{\emph{Arguments:}}{
\itemize{
\item{\strong{path:}}{
(\emph{character}) path of the file to create an \code{\link{Instance}}.
}
}
}
}
}
}
}

\seealso{
\code{\link{ExtractorEml}}, \code{\link{ExtractorSms}},
\code{\link{ExtractorTwtid}}, \code{\link{ExtractorYtbid}},
\code{\link{Instance}}
}
