% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SerialPipe.R
\docType{class}
\name{SerialPipe}
\alias{SerialPipe}
\title{Class implementing a default pipelining proccess.}
\usage{
SerialPipe
}
\description{
This \code{\link{SerialPipe}} class inherits from the
\code{\link{TypePipe}} class. Includes the \strong{pipeAll} method which
provides a default pipelining implementation.
}
\section{Constructor}{

\code{SerialPipe$new()}
}

\section{Details}{

This class uses the default implementation provided by its parent class
(\code{\link{TypePipe}}) and handles all the unexpected errors that may
appear during the execution of the pipelining process.
}

\section{Inherit}{

This class inherits from \code{\link{TypePipe}} and implements the
\code{pipeAll} abstract function.
}

\section{Methods}{

\itemize{
\item{\bold{pipeAll:}}{
implementation of the pipeAll function provided by the super class
(\code{\link{TypePipe}}). In this case, the pipeAll function is called
so that the default pipelining is executed.
\itemize{
\item{\emph{Usage:}}{
\code{pipeAll(instance)}
}
\item{\emph{Value:}}{
The preprocessed \code{\link{Instance}}.
}
\item{\emph{Arguments:}}{
\itemize{
\item{\strong{instance:}}{
(\emph{Instance}) The \code{\link{Instance}} that is going to be processed.
}
}
}
}
}
}
}

\seealso{
\code{\link{Instance}}, \code{\link{TypePipe}}
}
