% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TypePipe.R
\docType{class}
\name{TypePipe}
\alias{TypePipe}
\title{Absctract super class implementing the pipelining proccess.}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
TypePipe
}
\description{
Class to establish the flow of Pipes.
}
\section{Constructor}{

\code{TypePipe$new()}
}

\section{Details}{

The default flow is:

\preformatted{
instance \%>I\%

  TargetAssigningPipe$new()$pipe() \%>I\%

  StoreFileExtPipe$new()$pipe() \%>I\%

  GuessDatePipe$new()$pipe() \%>I\%

  File2Pipe$new()$pipe() \%>I\%

  MeasureLengthPipe$new()$pipe("length_before_cleaning_text") \%>I\%

  FindUserNamePipe$new()$pipe() \%>I\%

  FindHashtagPipe$new()$pipe() \%>I\%

  FindUrlPipe$new()$pipe() \%>I\%

  FindEmoticonPipe$new()$pipe() \%>I\%

  FindEmojiPipe$new()$pipe() \%>I\%

  GuessLanguagePipe$new()$pipe() \%>I\%

  ContractionPipe$new()$pipe() \%>I\%

  AbbreviationPipe$new()$pipe() \%>I\%

  SlangPipe$new()$pipe() \%>I\%

  ToLowerCasePipe$new()$pipe() \%>I\%

  InterjectionPipe$new()$pipe() \%>I\%

  StopWordPipe$new()$pipe() \%>I\%

  MeasureLengthPipe$new()$pipe("length_after_cleaning_text") \%>I\%

  TeeCSVPipe$new()$pipe()
}
}

\section{Methods}{

\itemize{
\item{\bold{pipeAll:}}{
function where is implemented the flow of the pipes.
\itemize{
\item{\emph{Usage:}}{
\code{pipeAll(instance)}
}
\item{\emph{Value:}}{
the preprocessed \code{\link{Instance}}.
}
\item{\emph{Arguments:}}{
\itemize{
\item{\strong{instance:}}{
(\emph{Instance}) the \code{\link{Instance}} that is going to be processed.
}
}
}
}
}
}
}

\seealso{
\code{\link{Instance}}, \code{\link{SerialPipe}}
}
