% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TypePipe.R
\docType{class}
\name{TypePipe}
\alias{TypePipe}
\title{Absctract super class implementing the pipelining proccess.}
\usage{
TypePipe
}
\description{
Abstract super class to establish the flow of Pipes.
}
\section{Constructor}{

\code{TypePipe$new()}
}

\section{Details}{

The default flow is:

\preformatted{
instance \%>I\%

  TargetAssigningPipe$new() \%>I\%

  StoreFileExtPipe$new() \%>I\%

  GuessDatePipe$new() \%>I\%

  File2Pipe$new() \%>I\%

  MeasureLengthPipe$new(propertyName = "length_before_cleaning_text") \%>I\%

  FindUserNamePipe$new() \%>I\%

  FindHashtagPipe$new() \%>I\%

  FindUrlPipe$new() \%>I\%

  FindEmoticonPipe$new() \%>I\%

  FindEmojiPipe$new() \%>I\%

  GuessLanguagePipe$new() \%>I\%

  ContractionPipe$new() \%>I\%

  AbbreviationPipe$new() \%>I\%

  SlangPipe$new() \%>I\%

  ToLowerCasePipe$new() \%>I\%

  InterjectionPipe$new() \%>I\%

  StopWordPipe$new() \%>I\%

  MeasureLengthPipe$new(propertyName = "length_after_cleaning_text") \%>I\%

  TeeCSVPipe$new()
}
}

\section{Methods}{

\itemize{
\item{\bold{pipeAll:}}{
function where is implemented the flow of the pipes.
\itemize{
\item{\emph{Usage:}}{
\code{pipeAll(instance)}
}
\item{\emph{Value:}}{
the preprocessed \code{\link{Instance}}.
}
\item{\emph{Arguments:}}{
\itemize{
\item{\strong{instance:}}{
(\emph{Instance}) the \code{\link{Instance}} that is going to be processed.
}
}
}
}
}
}
}

\seealso{
\code{\link{Instance}}, \code{\link{SerialPipe}},
         \code{\link{PipeGeneric}}, \code{\link{\%>I\%}}
}
